pyproject.toml
[tool.poetry]
name = "form"
version = "0.1.0"
description = ""
authors = ["EthanThatOneKid <31261035+EthanThatOneKid@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
mistune = "^3.0.1"
requests = "^2.31.0"
pyjwt = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
